feature: 添加新的 APIGW 接口
