+++ gcloud/apigw/views/__init__.py
@@ -1,0 +1,42 @@
+# -*- coding: utf-8 -*-
+"""
+Tencent is pleased to support the open source community by making 蓝鲸智云PaaS平台社区版 (BlueKing PaaS Community
+Edition) available.
+Copyright (C) 2017-2019 THL A29 Limited, a Tencent company. All rights reserved.
+Licensed under the MIT License (the "License"); you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+http://opensource.org/licenses/MIT
+Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+specific language governing permissions and limitations under the License.
+"""
+
+
+from .create_periodic_task import create_periodic_task  # noqa
+from .create_task import create_task  # noqa
+from .fast_create_task import fast_create_task  # noqa
+from .get_common_template_info import get_common_template_info  # noqa
+from .get_common_template_list import get_common_template_list  # noqa
+from .get_periodic_task_info import get_periodic_task_info  # noqa
+from .get_periodic_task_list import get_periodic_task_list  # noqa
+from .get_plugin_list import get_plugin_list  # noqa
+from .get_task_detail import get_task_detail  # noqa
+from .get_task_node_detail import get_task_node_detail  # noqa
+from .get_task_status import get_task_status  # noqa
+from .get_template_info import get_template_info  # noqa
+from .get_template_list import get_template_list  # noqa
+from .import_common_template import import_common_template  # noqa
+from .modify_constants_for_periodic_task import modify_constants_for_periodic_task  # noqa
+from .modify_cron_for_periodic_task import modify_cron_for_periodic_task  # noqa
+from .node_callback import node_callback  # noqa
+from .operate_task import operate_task  # noqa
+from .preview_task_tree import preview_task_tree  # noqa
+from .query_task_count import query_task_count  # noqa
+from .set_periodic_task_enabled import set_periodic_task_enabled  # noqa
+from .start_task import start_task  # noqa
+from .get_user_project_list import get_user_project_list  # noqa
+from .get_user_project_detail import get_user_project_detail  # noqa
+from .get_template_schemes import get_template_schemes  # noqa
+from .preview_task_tree import preview_task_tree  # noqa
+from .get_task_node_data import get_task_node_data  # noqa
+from .operate_node import operate_node  # noqa
