+++ gcloud/apigw/views/start_task.py
@@ -1,0 +1,48 @@
+# -*- coding: utf-8 -*-
+"""
+Tencent is pleased to support the open source community by making 蓝鲸智云PaaS平台社区版 (BlueKing PaaS Community
+Edition) available.
+Copyright (C) 2017-2019 THL A29 Limited, a Tencent company. All rights reserved.
+Licensed under the MIT License (the "License"); you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+http://opensource.org/licenses/MIT
+Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+specific language governing permissions and limitations under the License.
+"""
+
+
+from django.http import JsonResponse
+from django.views.decorators.csrf import csrf_exempt
+from django.views.decorators.http import require_POST
+
+from blueapps.account.decorators import login_exempt
+from gcloud.apigw.decorators import api_verify_perms
+from gcloud.apigw.decorators import mark_request_whether_is_trust
+from gcloud.apigw.decorators import project_inject
+from gcloud.taskflow3.models import TaskFlowInstance
+from gcloud.taskflow3.permissions import taskflow_resource
+
+try:
+    from bkoauth.decorators import apigw_required
+except ImportError:
+    from packages.bkoauth.decorators import apigw_required
+
+
+@login_exempt
+@csrf_exempt
+@require_POST
+@apigw_required
+@mark_request_whether_is_trust
+@project_inject
+@api_verify_perms(
+    taskflow_resource,
+    [taskflow_resource.actions.operate],
+    get_kwargs={'task_id': 'id', 'project_id': 'project_id'}
+)
+def start_task(request, task_id, project_id):
+    username = request.user.username
+    project = request.project
+    task = TaskFlowInstance.objects.get(pk=task_id, project_id=project.id)
+    ctx = task.task_action('start', username)
+    return JsonResponse(ctx)
