+++ gcloud/core/utils/__init__.py
@@ -36,7 +36,8 @@ ver_exports = [
     'convert_group_name',
     'convert_readable_username',
     'get_user_business_list',
-    'get_all_business_list'
+    'get_all_business_list',
+    'get_user_business_detail'
 ]
 for func in ver_exports:
     locals()[func] = getattr(ver_utils, func)
