+++ gcloud/taskflow3/api.py
@@ -39,6 +39,7 @@ from gcloud.commons.template.models import CommonTemplate
 from gcloud.tasktmpl3.models import TaskTemplate
 from gcloud.taskflow3.context import TaskContext
 from gcloud.contrib.analysis.analyse_items import task_flow_instance
+from gcloud.taskflow3.utils import preview_template_tree
 
 logger = logging.getLogger("root")
 get_client_by_user = settings.ESB_GET_CLIENT_BY_USER
@@ -265,23 +266,24 @@ def preview_task_tree(request, project_id):
     template_source = request.POST.get('template_source', PROJECT)
     template_id = request.POST.get('template_id')
     version = request.POST.get('version')
-    if template_source == PROJECT:
-        template = TaskTemplate.objects.get(pk=template_id, is_deleted=False, project_id=project_id)
-    else:
-        template = CommonTemplate.objects.get(pk=template_id, is_deleted=False)
     exclude_task_nodes_id = json.loads(request.POST.get('exclude_task_nodes_id', '[]'))
-    pipeline_tree = template.get_pipeline_tree_by_version(version)
-    template_constants = copy.deepcopy(pipeline_tree['constants'])
+
     try:
-        TaskFlowInstance.objects.preview_pipeline_tree_exclude_task_nodes(pipeline_tree, exclude_task_nodes_id)
+        data = preview_template_tree(
+            project_id,
+            template_source,
+            template_id,
+            version,
+            exclude_task_nodes_id
+        )
     except Exception as e:
-        logger.exception(e)
-        return JsonResponse({'result': False, 'message': str(e)})
-    constants_not_referred = {key: value for key, value in list(template_constants.items())
-                              if key not in pipeline_tree['constants']}
+        err_msg = 'preview_template_tree fail: {}'.format(e)
+        logger.exception(err_msg)
+        return JsonResponse({'result': False, 'message': err_msg})
+
     return JsonResponse({
         'result': True,
-        'data': {'pipeline_tree': pipeline_tree, 'constants_not_referred': constants_not_referred}
+        'data': data
     })
